Add Rake task to fix safe2pee encoding problems.
The seed data from safe2pee is a confusion of encodings, more easily
noticed where I live in Montréal. This Rake task applies
heuristically-determined transformations to fix legacy encoding errors
carried over from safe2pee.

It includes a "dry run" flag to allow you to preview the results of the
transformation before applying them to the database. This need only be
run once on the production database, but developers can also use it to
massage the seed data when running locally.