Use idiomatic syntax for rendering list of bathrooms
The existing code used a for loop with some item HTML in the primary template and
some in a partial.

I moved all of the repeated code into the partial, and then renamed the partial so
collection shorthand could be used instead of looping over a partial render call.

For more information on this shorthand, see section 3.4.5 of
http://guides.rubyonrails.org/layouts_and_rendering.html