Fix map rendering
Map rendering was broken because the rendering javascript could not find the variables that specify which images are used to indicate where the user is and where bathrooms are.

Those variable declarations were being made in the splash layout but not the application layout. I removed them from the splash layout, which does not need them. The only pages that show maps are in the bathrooms controller, so I created a new map partial in that directory to store shared map-related code and included that partial in the index and show views of that controller.