first commit - filter bar and map view redesign
in the process, had to create a wrapper div called constrained that
pushes certain elements inwards, to be used for things that one doesn’t
want lined up against the edge. To incorporate this so that the map
could be full width, I had to put this class around areas such as forms.

Known bugs - 1 ) The distance away column on the right of the bathroom
item is forced down by something. 2) The contact us page is broken by
constrained tag. I need to put it in.