DRY: mv rule from new.scss to formsubmit.html.haml
This commit moves a rule from new.scss into formsubmit.html.haml;
It removes a custom rule from new.css, which applied
margin-bottom: 20px to elements with class .nearby-container,
and implements identical styling via HAML.
Identical styling in HAML is achieved with the .footroom style,
which was already baked into the project.

I notice this rule spaces things out vertically, even when the
nearby results aren't showing. Oh well. It still looks good.
But it's getting to be a lot of vertical space, so if it needs
to be toned down a bit, I would understand.