Rebind map and buttons when getting new form
Clicking the guess button replaces the current form element with
a new one, which means the NewRestroomForm object's references get
outdated. The references, like @_map and @_previewButton, refer to
the elements in the old form. Operating on them don't do anything
to the new form. Calling _rebind binds the references to the new
elements in the new form.