config: Merge config file changes from Rails 5.1.4
Anything that would have changed functionality
has been manually commented out.

If we want to move to the new Rails 5.1 defaults,
we are going to have to research what these new defaults do,
and possibly switch to them and test that things work
the way we want them to.

This commit simply makes sure we have all the new config files,
but as mentioned above, everything new is turned off for now.

We may want to delete some of these (newly added) files
if we aren't going to use them.