Allow edit_id to be submitted to create by form
Add edit_id to permitted params and add it as a hidden field in the
form, which allows it to be submitted. The edit_id set in the 'new'
method only affects the value of the field in the form, and not the
edit_id for the @restroom in the create method, since they refer
to different objects.