Contact page: Fix h1 heading
h1 heading now properly renders with the Restroom listing's name,
when requesting contact from an existing restroom page.

(Translating this heading with one translation string won't work,
because dynamic Ruby content won't render properly
in a translation string.
In this case, the dynamic content is the name of the Restroom entry.)