simple_form.fr.yml: Revert key names to English
The quoted key names "yes" and "no" were changed to
"oui" and "non" in PR #450.

This changes them back to English "yes" and "no".

(Key names generally should not be translated,
and only the corresponding values need to be updated.)

For details about certain key names which need to be in quotes,
see the comments in: `config/locales/en/en.yml`