Release 1.12.0 (#481)
* Fix map not fading out after clicking 'List View'

This was a bug from my PR that refactored the map toggle.

* Hide changing table icon in minimal lists

For example, in a list of nearby restrooms, both the gender neutral
and accessibility icons are hidden while changing table icon still
shows. Fixed by this commit.

* added Polish translation

* fixed encoding

* style

* tells users that the text message service is unavailable in Poland

* fixed encoding again

* config/application.rb: Add Polish locale (pl)

This activates the Polish translations from #476.

The app should now show in Polish for browsers with "pl"
in a preferred position of their Accept_Language headers.