Empty commit for attribution
This is to give credit to @ashishwadekar for reviewing the preceding Pull Request.

(Accidentally added an extra ".com" to the email address
in the previous merge commit, so attribution on GitHub didn't work there.
Unable to fix that commit with a "git commit --amend",
due to force pushes being disabled on develop branch.)

Co-authored-by: Ashish Wadekar <ashishwadekar@users.noreply.github.com>