spec: Fix a path construction, test passes now
Passing the whole array of restroom data was causing the whole
array to be erroneously interpreted as if it was the :locale
prefix in the path.

Explicitly pass :id, and only :id,
to the `restroom_path` route helper for this test.

(We only need the :id from the newly-constructed test restroom
in order to visit the correct path. Other restroom data isn't
needed here anyhow.)

Co-authored-by: Mikena Wood <mi-wood@users.noreply.github.com>