Dynamically compose About and Business FAQ links
Dynamically compose the "Business FAQ" link on the "About" page
and the "About Page" link on the "Business FAQ" page
using Rails link helpers, so that the locale (and any other
default URL parameters) are automatically added to these links.

In other words, "un-hard-code" these links.

(In this commit, updated the Haml source of the About and Business FAQ
pages, and only the English internationalization (I18n) strings.)