deps: Resolve ansi-html to ansi-html-community
Pin to an exact version, 0.0.8, which I have personally verified
only contains the desired change, plus package.json metadata changes
and updates to README.md.

(Pinning means we don't have to trust the independent GitHub user
who published the fork, just the content of the patch.)

One can do the following commands to download and inspect the contents
of the packages themselves, as hosted on the npm package registry...
- `npm pack ansi-html@0.0.7`
- `npm pack ansi-html-community@0.0.8`
Then extract the tarballs and compare the content.

See the following for details of the fork and why it was made:
- github[dot]com/Tjatse/ansi-html/issues/19
- github[dot]com/Tjatse/ansi-html/issues/19#issuecomment-913119841