Docker: Update NodeJS version to 16.x
This is the new default over at Heroku.

We should start developing and testing against this version of Node,
in order to better match how our code will run in production/
on the real site.

(Reminder: We can always specify a version for production,
if we need to, with the "engines" field of the package.json file.)

See: https://devcenter.heroku.com/articles/nodejs-support#specifying-a-node-js-version