CI: Try using Ubuntu Jammy (22.04) in .travis.yml
There's some issue with the apt install step in the Dockerfile,
but only in CI. It has no issues for me locally.

Try using a newer base image for the CI. Maybe this has newer Docker,
which could avoid weird errors?

Currently implicitly uses Ubuntu Xenial (16.04!) which is quite old!